product create request updated

+++ app/Http/Requests/Products/CreateRequest.php
@@ -32,7 +32,7 @@ class CreateRequest extends FormRequest
             'dft' => ['required', 'integer', 'min:15'],
             'dry_to_touch' => ['required', 'integer', 'min:0'],
             'dry_to_handle' => ['required', 'integer', 'min:0'],
-            'min_int' => ['required', 'integer', 'min:0'],
+            'min_int' => ['required', 'numeric', 'min:0'],
             'max_int' => ['required', 'min:-1', 'numeric'],
             'min_temp' => ['required', 'integer', 'min:-20'],
             'max_service_temp' => ['required', 'integer', 'min:0'],
